pyproject.toml
[tool.poetry]
name = "lostindiffusion"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.0 <3.11.0"
fastapi = "^0.87.0"
jinja2 = "^3.1.2"
python-dotenv = "^0.21.0"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
diffusers = "^0.7.2"
transformers = "^4.24.0"
scipy = "^1.9.3"
ftfy = "^6.1.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"
poethepoet = "^0.16.4"


[tool.poe.tasks]
torch = "pip install torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu117"
accelerate = "pip install accelerate"

[[tool.poetry.source]]
name = "nvidia"
url = "https://pypi.ngc.nvidia.com"
default = false
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
